chore(routes): add root routes grouping to allow empty root layout
