+++ src/routes/(scores)/+layout.server.ts
@@ -1,0 +1,6 @@
+import type { LayoutServerLoad } from '../$types';
+
+export const load: LayoutServerLoad = async ({ parent }) => {
+  var parentLayoutData = await parent();
+  return parentLayoutData;
+};
