+++ src/routes/(scores)/[[lang=lang]]/+page.server.ts
