+++ src/routes/(scores)/[[lang=lang]]/+page.svelte
