+++ src/routes/(scores)/[[lang=lang]]/[competitions=competitions]/+page.server.ts
