+++ src/routes/(scores)/[[lang=lang]]/[competitions=competitions]/+page.svelte
