+++ src/routes/(scores)/[[lang=lang]]/[competitions=competitions]/[...competition_fill]/+page.server.ts
