+++ src/routes/(scores)/[[lang=lang]]/[player=player]/[...player_fill]/+page.svelte
