+++ src/routes/(scores)/[[lang=lang]]/[sport]/+page.ts
