+++ src/routes/(scores)/[[lang=lang]]/[sport]/[country]/+page.ts
