+++ src/routes/(scores)/[[lang=lang]]/[sport]/[country]/[league_name]/+page.server.ts
