+++ src/routes/(scores)/[[lang=lang]]/[sport]/[fixture=fixture]/+page.server.ts
