+++ src/routes/(scores)/u/[view]/[lang=lang]/+page.svelte
